Introduce LockItem and AccountLocks
