+++ packages/ui/src/app/pages/Profile/components/AccountItem.tsx
@@ -3,180 +3,16 @@ import React from 'react'
 import styled from 'styled-components'
 
 import { AccountInfo } from '@/accounts/components/AccountInfo'
+import { AccountLocks } from '@/accounts/components/AccountLocks'
 import { TransferButton } from '@/accounts/components/TransferButton'
 import { useBalance } from '@/accounts/hooks/useBalance'
-import { Account, BalanceLock, LockType } from '@/accounts/types'
-import { ButtonGhost } from '@/common/components/buttons'
+import { Account } from '@/accounts/types'
+import { DetailsItemVote, LockItem } from '@/app/pages/Profile/components/LockItem'
 import { DropDownButton, DropDownToggle } from '@/common/components/buttons/DropDownToggle'
-import { Arrow } from '@/common/components/icons'
-import { BoxIcon } from '@/common/components/icons/BoxIcon'
-import { EnvelopeIcon } from '@/common/components/icons/EnvelopeIcon'
-import { LabelIcon } from '@/common/components/icons/LabelIcon'
-import { VoteIcon } from '@/common/components/icons/VoteIcon'
-import { TextSmall, TextMedium, TokenValue } from '@/common/components/typography'
+import { TextMedium, TokenValue } from '@/common/components/typography'
 import { Label } from '@/common/components/typography/Label'
 import { Sizes, Colors } from '@/common/constants'
 import { useToggle } from '@/common/hooks/useToggle'
-
-interface DetailsItemDataProps {
-  lock: BalanceLock
-}
-
-export const lockIcon = (type: LockType) => {
-  switch (type) {
-    case 'Voting':
-      return <LabelIcon />
-    case 'Council Candidate':
-      return <LabelIcon />
-    case 'Councilor':
-      return <LabelIcon />
-    case 'Validation':
-      return <LabelIcon />
-    case 'Nomination':
-      return <LabelIcon />
-    case 'Proposals':
-      return <LabelIcon />
-    case 'Storage Worker':
-      return <LabelIcon />
-    case 'Content Directory Worker':
-      return <LabelIcon />
-    case 'Forum Worker':
-      return <LabelIcon />
-    case 'Membership Worker':
-      return <LabelIcon />
-    case 'Invitation':
-      return <LabelIcon />
-    case 'Staking Candidate':
-      return <EnvelopeIcon />
-  }
-}
-
-export const LockItem = ({ lock }: DetailsItemDataProps) => {
-  const [isDropped, setDropped] = useToggle()
-
-  return (
-    <>
-      <AccountDetailsWrap>
-        <DetailsInfo>
-          {lockIcon(lock.type)}
-          <DetailsName>{lock.type}</DetailsName>
-        </DetailsInfo>
-        <TokenValue value={0} />
-        <TokenValue value={lock.amount} />
-        <TokenValue value={0} />
-        <TokenValue value={0} />
-        <DropDownButton onClick={setDropped} isDropped={isDropped} size="medium" />
-      </AccountDetailsWrap>
-      <DetailsDropDownToggle isDropped={isDropped}>
-        <DetailsDropDownColumn>
-          <DetailBox>
-            <DetailBoxTitle>Lock date:</DetailBoxTitle>
-            <TextMedium dark>01/07/2020, 10:00 am CET</TextMedium>
-            <DetailBlockNumber>
-              <BoxIcon />
-              <DetailBlockNumberText lighter>389,829 block</DetailBlockNumberText>
-            </DetailBlockNumber>
-          </DetailBox>
-          <DetailBox>
-            <DetailBoxTitle>Role:</DetailBoxTitle>
-            <TextMedium dark>Storage Provider</TextMedium>
-          </DetailBox>
-        </DetailsDropDownColumn>
-        <DetailsDropDownColumn>
-          <DetailsButton size="small">
-            Application preview
-            <ButtonArrow direction="right" />
-          </DetailsButton>
-          <DetailsButton size="small">
-            Opening preview
-            <ButtonArrow direction="right" />
-          </DetailsButton>
-        </DetailsDropDownColumn>
-      </DetailsDropDownToggle>
-    </>
-  )
-}
-
-interface DetailsItemLockDataProps {
-  account: Account
-}
-
-export const DetailsItemLock = ({ account }: DetailsItemLockDataProps) => {
-  const address = account.address
-  const balance = useBalance(address)
-
-  const [isDropped, setDropped] = useToggle()
-
-  return (
-    <>
-      <AccountDetailsWrap>
-        <DetailsInfo>
-          <EnvelopeIcon />
-          <DetailsName>Invitation lock</DetailsName>
-        </DetailsInfo>
-        <TokenValue value={balance?.total} />
-        <TokenValue value={balance?.locked} />
-        <TokenValue value={balance?.recoverable} />
-        <TokenValue value={balance?.transferable} />
-        <DropDownButton onClick={setDropped} isDropped={isDropped} size="medium" />
-      </AccountDetailsWrap>
-      <DetailsDropDownToggle isDropped={isDropped}>Is open</DetailsDropDownToggle>
-    </>
-  )
-}
-
-interface DetailsItemVoteDataProps {
-  account: Account
-}
-
-export const DetailsItemVote = ({ account }: DetailsItemVoteDataProps) => {
-  const address = account.address
-  const balance = useBalance(address)
-
-  const [isDropped, setDropped] = useToggle()
-
-  return (
-    <>
-      <AccountDetailsWrap>
-        <DetailsInfo>
-          <VoteIcon />
-          <DetailsName>Staking for voting</DetailsName>
-        </DetailsInfo>
-        <TokenValue value={balance?.total} />
-        <TokenValue value={balance?.locked} />
-        <TokenValue value={balance?.recoverable} />
-        <TokenValue value={balance?.transferable} />
-        <DropDownButton onClick={setDropped} isDropped={isDropped} size="medium" />
-      </AccountDetailsWrap>
-      <DetailsDropDownToggle isDropped={isDropped}>
-        <DetailsDropDownColumn>
-          <DetailBox>
-            <DetailBoxTitle>Lock date:</DetailBoxTitle>
-            <TextMedium dark>01/07/2020, 10:00 am CET</TextMedium>
-            <DetailBlockNumber>
-              <BoxIcon />
-              <DetailBlockNumberText lighter>389,829 block</DetailBlockNumberText>
-            </DetailBlockNumber>
-          </DetailBox>
-          <DetailBox>
-            <DetailBoxTitle>Role:</DetailBoxTitle>
-            <TextMedium dark>Storage Provider</TextMedium>
-          </DetailBox>
-        </DetailsDropDownColumn>
-        <DetailsDropDownColumn>
-          <DetailsButton size={'small'}>
-            Application preview
-            <ButtonArrow direction="right" />
-          </DetailsButton>
-          <DetailsButton size={'small'}>
-            Opening preview
-            <ButtonArrow direction="right" />
-          </DetailsButton>
-        </DetailsDropDownColumn>
-      </DetailsDropDownToggle>
-    </>
-  )
-}
 
 interface AccountItemDataProps {
   account: Account
@@ -189,20 +25,12 @@ export const AccountItem = ({ account }: AccountItemDataProps) => {
 
   const [isDropped, setDropped] = useToggle()
 
-  const displayLockIcons = () => {
-    if (!balance || !balance.locks.length) {
-      return null
-    }
-
-    return balance.locks.map((lock) => lockIcon(lock.type || 1))
-  }
-
   const displayLocks = () => {
     if (!balance || !balance.locks.length) {
       return <TextMedium light>No locks found.</TextMedium>
     }
 
-    return balance.locks.map((lock) => <LockItem key={lock.type + '_' + lock.amount.toString()} lock={lock} />)
+    return balance.locks.map((lock, index) => <LockItem key={index} lock={lock} />)
   }
 
   return (
@@ -213,7 +41,7 @@ export const AccountItem = ({ account }: AccountItemDataProps) => {
         <div>
           <TokenValue value={balance?.locked} />
           <br />
-          {displayLockIcons()}
+          <AccountLocks locks={balance?.locks} />
         </div>
         <TokenValue value={balance?.recoverable} />
         <TokenValue value={balance?.transferable} />
@@ -265,71 +93,4 @@ const StyledDropDown = styled(DropDownToggle)`
 
 const StyledLabel = styled(Label)`
   margin-bottom: 8px;
-`
-
-const AccountDetailsWrap = styled(AccountItemWrap)`
-  grid-template-columns: 260px repeat(4, 133px) 115px;
-  height: 46px;
-  padding: 3px 16px;
-  margin-top: 4px;
-  border-radius: 2px;
-  background-color: ${Colors.White};
-`
-
-const DetailsDropDownToggle = styled(DropDownToggle)`
-  display: flex;
-  align-items: flex-end;
-  flex-direction: row;
-  justify-content: space-between;
-  padding: 0 16px 16px 16px;
-  background-color: ${Colors.White};
-`
-
-const DetailsDropDownColumn = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  height: 100%;
-  width: 100%;
-
-  &:last-child {
-    display: flex;
-    justify-content: flex-end;
-  }
-`
-
-const DetailsInfo = styled.div`
-  display: flex;
-  width: 100%;
-`
-
-const DetailsName = styled.h6`
-  font-size: 14px;
-  font-weight: 700;
-  line-height: 20px;
-  color: ${Colors.Black[900]};
-  margin-left: 16px;
-`
-
-const DetailBox = styled.div``
-
-const DetailBoxTitle = styled(Label)`
-  margin-bottom: 8px;
-`
-
-const ButtonArrow = styled(Arrow)`
-  fill: ${Colors.Black[400]};
-`
-
-const DetailsButton = styled(ButtonGhost)`
-  margin-left: 10px;
-`
-
-const DetailBlockNumber = styled.div`
-  display: flex;
-  align-items: center;
-  margin-top: 4px;
-`
-
-const DetailBlockNumberText = styled(TextSmall)`
-  margin-left: 4px;
 `
